REMOVE T-MOBILE THEME ENGINE (1/3)
As it stands right now, this is only functional on 2-3 devices and for some reason, is not on about 4-5 devices.

I've never seen anything like this before but it is obvious that there are some conflicts that need to be fixed.

With that said, I want it out of DU for now since is obviously not stable and until we can fix the conflicts.

Change-Id: I1a7bd3ebbc70d9f3b80d14a50ec03217abff909a