Move CPU info to the top left / add transparency (see notes)
So it seems there was an issue where we picked it from someone (AOSB) who was kanging it from
the original author (Alexander) and we (DU) removed the commit in question.

I talked to the original author and we figured things out so I'm adding this back in.

Oh and see pic below
http://i.imgur.com/Y75bmRL.png

Change-Id: I022386ab04d5ee19999a7dca64580064694e9baf