Frameworks: SlimRecents: go more inline with AOSP and add menu + new googly shortcuts
This commit changes the longpress behaviour. Like on stock aosp longpress opens the menu with this options

1. App info (like stock)
2. Play Store link to the app
3. Amazon store link to the app

Note: the store links are only shown if the app has the info that it was installed by one of this
stores. So stock apps (installed by system) or sideloaded apps do not show this link.

Change-Id: I29769f48ce6073acb6105f5590b93c671d6fc0d4