[3/4] Finish fixing Zygote descriptor leakage problem
In order to prevent Zygote descriptors from leaking into the child
environment, they should be closed by the forked-off child process
before the child switches to the application UID.  These changes close
the descriptors via dup2(), substituting a descriptor open to
/dev/null in their place; this allows the Zygote Java code to close
the FileDescriptor objects cleanly.

This is a multi-project change: dalvik, art, libcore, frameworks/base,
and external/sepolicy are affected.  The CLs need to be approved
together, lest the build break or the software fail to boot.

Round 2: indent change

Also needed for this to compile and work fine: 

http://gerrit.dirtyunicorns.com/#/q/status:open+project:android_art,n,z
http://gerrit.dirtyunicorns.com/#/q/status:open+project:libcore,n,z and 
http://gerrit.dirtyunicorns.com/4726 or http://gerrit.dirtyunicorns.com/#/c/5031/



Bug: 12114500
Change-Id: I090402136a8a8b7d6aad6eb153026e85d7cf6ad3