Don't add common assets if files cannot be accessed
- If resApkPath does not exist, asset manaager will load the resources
from the theme's apk which end up overriding resources in the 0x7f
package id.  This causes apps, such as SystemUI, to crash endlessly.

The correct behavior is to not add common overlay assets if the
files that are being passed in as paths cannot be accessed.  The
same behavior is already used in AssetManager::addOverlayPath()

Change-Id: If5e90baeda5f07973f268060dcb6d73dae657603