SystemUI: Fix can't start camera via lock screen button
In keyguard secure mode, if start camera again, it will bring the existing
camera task to the front. But if its top activity is not CameraActivity,
and it can not hide keyguard, the device will look like no response.

Refer to KK's design, add new task and clear top flags for camera intent.

Change-Id: Ia9f6caa7b33242020f3c32fcd25c4261623ac36b
CRs-Fixed: 832604