Use Torch Service for Flashlight
Revert "Flashlight for Offscreen Gestures"

This reverts commit 36c1dcad4ce3980d40113d63e9a62e4886cd8665.

Implement system-wide Torch service

Allows Torch and any camera to coexist peacefully. When torch is active
and a camera is opened, the torch will be first shut down, and the
camera will be allowed to continue opening, instead of displaying
"Cannot connect to camera."

Also a squash of the following commits:

    add framework torch service

    Change-Id: I0c7a055b33fdebe697f23d207b20408e06eb74a8
    Signed-off-by: Roman Birg <roman@cyngn.com>
    Fully fix interaction between torch and camera usage.

    For doing this, two things are necessary:
    - The torch shutdown needs to be synchronous, so that the torch app has
      a chance to release the camera before the camera app tries to use it.
    - The torch service needs to track camera usage, so that it can notify
      the torch app if a camera is busy.

    Change-Id: I521091d255ff4c251d97a0e8d65c9d2a80b9dae4

    Fix camera usage tracking.

    When a Camera object is finalized, it notifies the torch service that
    it's no longer using the camera. This call used to not check whether the
    caller actually is still using the camera and called release() before,
    so depending on timing of the finalizer, it could happen that e.g. the
    camera app released the camera user record of the torch app, leading to
    the torch usage not being detected on next usage of the camera app.

    Fix that by checking whether the caller of onCameraClosed() actually
    owned the camera prior to releasing the record.
    Also fix synchronization between accesses.

    Change-Id: Ic01cbc3c89ab8d855e4613bfa1b2f548de919b73

    frameworks: extend TorchService to be able to handle sysfs torch config

    Some devices are still using a sysfs node to control the torch. Let's
    support those configurations and properly shut the torch down.

    Change-Id: I5fa1cccdcffad26dbfef4b2ee0eafe1218a5308b
    Signed-off-by: Roman Birg <roman@cyngn.com>

    frameworks: move TorchService to new services location

    Change-Id: I392f37dfefe41686388811069999eb031f7eb92c
    Signed-off-by: Roman Birg <roman@cyngn.com>

Change-Id: Id59df818b88dd01c4529e7b677faea9815de0486

Fix flashlight tile inconsistency

Toggling the torch via an external method (such as drawing V on
screen) would put the tile in a strange state and hide the tile.

Update the tile availability properly.

Change-Id: I1f0edcd8c0398f903fbaa31fc55a20962c0289ca

Fix up torch tile handling.

- Keep it in sync with other torch invocations (e.g. via gesture)
- Only treat torch as 'unavailable' if another client (camera app) uses
  the camera, but not if torch itself opened it.

Change-Id: I788c19bc73ad8554eede705d53c744cebaed9490

TorchService: improve state management

Always dispatch state when tearing torch down.

Change-Id: Ia2fbab421e552bf48f951625dd4eea48ace20363
Signed-off-by: Roman Birg <roman@cyngn.com>

Torch: remind user flashlight is still on

Post a notification which informs the user the flashlight is still
enabled. But only do this when the user turns the screen off *while* the
flashlight is still on, so they see it next time they use the device.
Tapping on the notification will disable the flashlight.

Change-Id: Ic708719135b91ba2177c525b1bd74b3ea0fbed7e

TorchService: close all opened cameras

When the torch is toggled on and then off very quickly, the camera can
still be opened and we never actually end up closing it.

Also reset the torch camera id when we tear the torch down.

Change-Id: I0175e8074edfc89e70f3278649d0ed8f69fb0a71
Signed-off-by: Roman Birg <roman@cyngn.com>

more fix for torch

Change-Id: Iabd2a6c0fd610bf9005b35be0c1d6f20d668369d

Squash of all Torch Sevice Commits

Change-Id: I89ccacc4026390bfb5acff1d6100222f5335acfa