Fix single action power menu issue
Currently if you set screenshot or say screen record or a single action in the power
menu, as soon as you longpress on the power button it triggers that action.

This may have been the intended way Google wanted things to work but it doesn't work for
us and in my opinion, is bad design.

Change-Id: Ib22a56dce747f81cfaa32ee162b98d92e8c3c2fb