Fix Bluetooth icon not properly refreshing
In the previous commit, a ContentObserver was being used and wasn't picking up
on the changes to the preference until the notification shade was toggled.

Borrowing heavily from the logic of c4d4dc585cc1581ce656b22ead3352a28609a8ae,
using a SettingsObserver fixes this and allows the toggle to happen in realtime.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Change-Id: I4b6c9d9f5984c8867e1746545a910cd3a60bcc13