Revert "Enable power save config of system apps [1/2]"
Reverting this because while this commit it gives us the ability to 
'optimize' system apps, it causes a huge problem. If the user does 
optimize the "Download Manager", it crashes the media provider while 
connecting the device via MTP or doing various other tasks. The little 
bit we gain from this 'power save' is just not worth the risk.

This reverts commit 3acf5c41fde1a4233da3b8867e01d1937bb54cf6.

Change-Id: I60e881c767fb7183ea1276d682d2fd500f45913b