Set Assistant UID of the current user
As per the new concurrent capture policy introduced in Android-10, on user 10 (Driver), when Voice recording is ongoing, if user says "Ok, Google", Audio policy shall forward the audio data to both record clients (Voice recorder and Assitant); However, Assistant gets silence, because Audio policy perceives Assitant as an ordinary record client. This is due to the wrong Assistant UID set in AudioSystem.java when user-10 is logged in.

This commit ensures to set the right Assistant Uid as that of the current user logged-in.

Change-Id: Id838a250c021cbd56fbe12dda9690d01cb9a9832
Tracked-On:
Test: manual - Triggered assistant during Voice recorder
Signed-off-by: Prashanth M, Shakthi <shakthi.prashanth.m@intel.com>
Co-authored-by: Prashanth M, Shakthi <shakthi.prashanth.m@intel.com>