QS customizer: Better string for 'reset' menu item
I know this is small and to some, maybe stupid but leaving it as 'reset' gives
the impression that it will reset the entire QS panel. Now that we have a few
customization options in place, this needed to change to let the user what this
menu item is really doing.

Change-Id: I10c985d0cf6554402e4bf50049a56d3bd6a38f05