Change default wallpaper name
Reason for this change is the same for why we changed the default wallpaper component name last year for the P3XL. Google keeps setting LWP's and default ones via overlay among other things. Instead of recompiling a vendor and having users flash this 'special' one every time they use DU, we figured this is best course of action.

Change-Id: Ia12e1b077ae08af3fb97c9504d672c4d103f9042